+++ qcm_interface/include/qcm_interface/model/radio.h
@@ -14,7 +14,7 @@ struct RadioRefer {
     QML_ANONYMOUS
 public:
     GADGET_PROPERTY_DEF(qcm::model::ItemId, itemId, id)
-    GADGET_PROPERTY_DEF(qint32, libraryId, libraryId)
+    GADGET_PROPERTY_DEF(qint64, libraryId, libraryId)
     GADGET_PROPERTY_DEF(QString, name, name)
     GADGET_PROPERTY_DEF(QString, picUrl, picUrl)
     std::strong_ordering operator<=>(const RadioRefer&) const = default;
